% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{auc}
\alias{auc}
\title{Area under the ROC curve by the rank statistic}
\usage{
auc(y, p)
}
\arguments{
\item{y}{Binary responses (0/1).}

\item{p}{Scores.}
}
\value{
AUC in \code{[0, 1]}.
}
\description{
Area under the ROC curve by the rank statistic
}
