% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world.R
\name{cell_xy}
\alias{cell_xy}
\title{Cell centres in kilometres}
\usage{
cell_xy(idx)
}
\arguments{
\item{idx}{Integer cell indices in \code{1:GRID^2}.}
}
\value{
A two column matrix of \code{x}, \code{y} coordinates (km).
}
\description{
Cell centres in kilometres
}
