% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world.R
\name{covariates}
\alias{covariates}
\title{Covariate matrix at a set of cells and a timestep}
\usage{
covariates(w, idx, t)
}
\arguments{
\item{w}{A world from \code{\link{make_world}}.}

\item{idx}{Integer cell indices.}

\item{t}{Timestep.}
}
\value{
A matrix with columns \code{depth}, \code{subst}, \code{wave}, \code{sst}.
}
\description{
Covariate matrix at a set of cells and a timestep
}
