% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{edf}
\alias{edf}
\title{Euclidean distance field block (E)}
\usage{
edf(query_xy, anchors)
}
\arguments{
\item{query_xy}{Query coordinates (km).}

\item{anchors}{Anchor coordinates, usually k-means centres of the training sites.}
}
\value{
A matrix with one column per anchor.
}
\description{
Euclidean distance field block (E)
}
