% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grf.R
\name{fftfreq}
\alias{fftfreq}
\title{Discrete Fourier frequencies}
\usage{
fftfreq(n)
}
\arguments{
\item{n}{Integer length.}
}
\value{
Numeric vector of length \code{n}.
}
\description{
Equivalent of \code{numpy.fft.fftfreq(n)} with unit sample spacing.
}
\keyword{internal}
