% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marineSim-package.R
\docType{package}
\name{marineSim-package}
\alias{marineSim}
\alias{marineSim-package}
\title{marineSim: the stRF simulation engine}
\description{
Functions are grouped as follows.
\itemize{
  \item World: \code{\link{matern_grf}}, \code{\link{make_world}},
    \code{\link{world_grid}}, \code{\link{cell_xy}}, \code{\link{covariates}}.
  \item Designs and datasets: \code{\link{sample_sites}},
    \code{\link{build_dataset}}, \code{\link{dataset_seed}}.
  \item Feature blocks: \code{\link{spatial_lags}}, \code{\link{edf}},
    \code{\link{nearest_train}}, \code{\link{dist_km}}.
  \item Models and scoring: \code{\link{score}}, \code{\link{glm_score}},
    \code{\link{auc}}.
  \item Attribution: \code{\link{required_subsets}}, \code{\link{skey}},
    \code{\link{shapley_row}}, \code{\link{add_shapley}}, \code{\link{boot_ci}},
    \code{\link{decomp_table}}.
  \item Extrapolation diagnostic: \code{\link{area_of_applicability}}.
  \item Constants: \code{\link{constants}}.
}
The analysis scripts in the repository's \code{analysis/} directory drive the
full benchmark, the auxiliary studies and the seagrass case study.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GITHUB-USER/marineSim}
  \item Report bugs at \url{https://github.com/GITHUB-USER/marineSim/issues}
}

}
\author{
\strong{Maintainer}: Ebenezer Afrifa-Yamoah \email{e.afrifayamoah@ecu.edu.au} (\href{https://orcid.org/0000-0003-1741-9249}{ORCID})

Authors:
\itemize{
  \item Francky Fouedjio
  \item Ute Mueller
  \item Kathryn McMahon
  \item Glenn A. Hyndes
}

}
\keyword{internal}
