% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{nearest_train}
\alias{nearest_train}
\title{Index of the nearest training site for each query point}
\usage{
nearest_train(train_xy, query_xy, exclude_self = FALSE)
}
\arguments{
\item{train_xy}{Training coordinates (km).}

\item{query_xy}{Query coordinates (km).}

\item{exclude_self}{Exclude the zero distance match (query set is the training set).}
}
\value{
Integer vector of training row indices.
}
\description{
Index of the nearest training site for each query point
}
